# packages/api-types/openapi.json
{
  "components": {
    "schemas": {
      "APIError": {
        "additionalProperties": false,
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorBody"
          }
        },
        "required": [
          "error"
        ],
        "type": "object"
      },
...
      "ErrorBody": {
        "additionalProperties": false,
        "properties": {
          "code": {
            "description": "Stable, machine-readable error code.",
            "examples": [
              "invalid_request"
            ],
            "type": "string"
          },
          "details": {
            "description": "Per-field problems for validation errors.",
            "items": {
              "$ref": "#/components/schemas/ErrorDetail"
            },
            "type": "array"
          },
          "message": {
            "description": "Human-readable explanation, safe to show to developers.",
            "examples": [
              "The request body is not valid JSON."
            ],
            "type": "string"
          },
          "request_id": {
            "description": "Include this when contacting support or searching logs.",
            "examples": [
              "req_01j9tq4m2xk3v8c7e5r2n0w6yb"
            ],
            "type": "string"
          }
        },
        "required": [
          "code",
          "message"
        ],
        "type": "object"
      },
      "ErrorDetail": {
        "additionalProperties": false,
        "properties": {
          "location": {
            "description": "Where the error occurred, e.g. 'body.items[3].tags' or 'path.thing-id'",
            "type": "string"
          },
          "message": {
            "description": "Error message text",
            "type": "string"
          },
          "value": {
            "description": "The value at the given location"
          }
        },
        "type": "object"
      },
...
      "LoginInputBody": {
        "additionalProperties": false,
        "properties": {
          "email": {
            "examples": [
              "ada@example.com"
            ],
            "format": "email",
            "maxLength": 254,
            "type": "string"
          },
          "password": {
            "maxLength": 512,
            "minLength": 1,
            "type": "string"
          }
        },
        "required": [
          "email",
          "password"
        ],
        "type": "object"
      },
      "MeResponse": {
        "additionalProperties": false,
        "properties": {
          "organizations": {
            "items": {
              "$ref": "#/components/schemas/Organization"
            },
            "type": "array"
          },
          "user": {
            "$ref": "#/components/schemas/User"
          }
        },
        "required": [
          "user",
          "organizations"
        ],
        "type": "object"
      },
...
      "Organization": {
        "additionalProperties": false,
        "properties": {
          "created_at": {
            "format": "date-time",
            "type": "string"
          },
          "id": {
            "examples": [
              "org_01j9tq4m2xk3v8c7e5r2n0w6yb"
            ],
            "type": "string"
          },
          "name": {
            "examples": [
              "Acme"
            ],
            "type": "string"
          },
          "role": {
            "description": "Your role in this organization.",
            "enum": [
              "owner",
              "admin",
              "member"
            ],
            "type": "string"
          },
          "slug": {
            "examples": [
              "acme"
            ],
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "slug",
          "role",
          "created_at"
        ],
        "type": "object"
      },
...
      "User": {
        "additionalProperties": false,
        "properties": {
          "created_at": {
            "format": "date-time",
            "type": "string"
          },
          "email": {
            "examples": [
              "ada@example.com"
            ],
            "type": "string"
          },
          "email_verified": {
            "type": "boolean"
          },
          "id": {
            "examples": [
              "usr_01j9tq4m2xk3v8c7e5r2n0w6yb"
            ],
            "type": "string"
          },
          "name": {
            "examples": [
              "Ada Lovelace"
            ],
            "type": "string"
          },
          "operator": {
            "description": "Whether you operate this Bridge instance and can see System health.",
            "type": "boolean"
          }
        },
        "required": [
          "id",
          "operator",
          "email",
          "name",
          "email_verified",
          "created_at"
        ],
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "apiKey": {
        "bearerFormat": "bk_live_… or bk_test_…",
        "description": "Project API key. Create one in the dashboard under API keys. Test keys never send real SMS.",
        "scheme": "bearer",
        "type": "http"
      },
      "deviceCredential": {
        "bearerFormat": "bd_…",
        "description": "Device credential issued to the Android gateway when it pairs.",
        "scheme": "bearer",
        "type": "http"
      },
      "session": {
        "description": "Dashboard session cookie, set by `POST /v1/auth/login`.",
        "in": "cookie",
        "name": "bridge_session",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "description": "Open-source infrastructure for SMS and phone verification. Developer endpoints authenticate with an API key (`Authorization: Bearer bk_live_…`). Dashboard endpoints use a session cookie.",
    "license": {
      "identifier": "AGPL-3.0-only",
      "name": "AGPL-3.0-only"
    },
    "title": "Bridge API",
    "version": "dev"
  },
  "openapi": "3.1.0",
  "paths": {
    "/v1/auth/login": {
      "post": {
        "operationId": "login",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginInputBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MeResponse"
                }
              }
            },
            "description": "OK",
            "headers": {
              "Set-Cookie": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIError"
                }
              }
            },
            "description": "Unauthorized"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIError"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIError"
                }
              }
            },
            "description": "Too Many Requests"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIError"
                }
              }
            },
            "description": "Internal Server Error"
          }
        },
        "summary": "Sign in",
        "tags": [
          "Auth"
        ]
      }
    },
...
  },
  "servers": [
    {
      "url": "http://localhost:8080"
    }
  ],
...
}